fix: rename the file typo  (#3271)
fix: rename the file typo
