+++ src/batch/src/task/mod.rs
@@ -14,7 +14,7 @@
 
 pub use context::*;
 pub use env::*;
-pub use task_::*;
+pub use task_execution::*;
 pub use task_manager::*;
 
 mod broadcast_channel;
@@ -23,5 +23,5 @@ mod context;
 mod env;
 mod fifo_channel;
 mod hash_shuffle_channel;
-mod task_;
+mod task_execution;
 mod task_manager;
